Fixed Circus Train
Moved the functionality of the, now deleted, PlatformManager to the Train.
The train now receives animals which are organised into lists and can then be placed into wagons.

Added the unit tests necesary to test the algorithm meets the requirements (see WagonTests for these)

Removed the PlatformManager and its tests as it has become obsolete.